fix: add russian grade-7 route redirect to cheatsheet
